+++ lib/src/utils.dart
@@ -8,9 +8,7 @@ Locale localeFromString(String localeString) {
       return Locale(localeList.first, localeList.last);
     case 3:
       return Locale.fromSubtags(
-          languageCode: localeList.first,
-          scriptCode: localeList[1],
-          countryCode: localeList.last);
+          languageCode: localeList.first, scriptCode: localeList[1], countryCode: localeList.last);
     default:
       return Locale(localeList.first);
   }
@@ -19,19 +17,4 @@ Locale localeFromString(String localeString) {
 /// Convert [locale] to Srting with custom [separator]
 String localeToString(Locale locale, {String separator = '_'}) {
   return locale.toString().split('_').join(separator);
-}
-
-/// Emit a [info] log event
-void printInfo(String info) {
-  print('\u001b[32mEasy Localization: $info\u001b[0m');
-}
-
-/// Emit a [warning] log event
-void printWarning(String warning) {
-  print('\u001B[34m[WARNING] Easy Localization: $warning\u001b[0m');
-}
-
-/// Emit a [error] log event
-void printError(String error) {
-  print('\u001b[31m[ERROR] Easy Localization: $error\u001b[0m');
 }